// src/lib/api/generated/model/registrationDto.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * BASE API
 * OpenAPI spec version: 3.0
 */

export interface RegistrationDto {
  fullName: string;
  /** Tên đăng nhập: chỉ chữ thường và số (hoặc số điện thoại), không ký tự đặc biệt, 4–30 ký tự */
  username: string;
  password: string;
  /** Email nhận mã OTP xác thực */
  email: string;
  /** Số điện thoại (tuỳ chọn) */
  phone?: string;
}