// src/lib/api/generated/model/userResDto.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * BASE API
 * OpenAPI spec version: 3.0
 */
import type { EUserStatus } from './eUserStatus';
import type { EUserType } from './eUserType';

export interface UserResDto {
  id: number;
  phone?: string;
  username?: string;
  fullName?: string;
  email?: string;
  avatar?: string;
  address?: string;
  status?: EUserStatus;
  type?: EUserType;
  createdAt?: string;
  lastLoginAt?: string;
}